+++ sample/src/desktopMain/kotlin/com/github/panpf/sketch/sample/ui/setting/AppSettingsDialog.desktop.kt
@@ -5,6 +5,18 @@ import com.github.panpf.sketch.sample.EventBus
 
 actual fun platformMakeDecodeMenuList(appSettings: AppSettings): List<SettingItem> = emptyList()
 
+actual fun platformAnimatedMenuList(appSettings: AppSettings): List<SettingItem> = buildList {
+    add(
+        SwitchSettingItem(
+            title = "Cache Decode Timeout Frame",
+            desc = null,
+            state = appSettings.cacheDecodeTimeoutFrame,
+        )
+    )
+}
+
+
+
 actual fun platformMakeOtherMenuList(appSettings: AppSettings): List<SettingItem> = buildList {
     add(
         DropdownSettingItem(