+++ sketch-animated/src/nonAndroidMain/kotlin/com/github/panpf/sketch/decode/internal/SkiaAnimatedDecoder.kt
@@ -1,14 +1,15 @@
 package com.github.panpf.sketch.decode.internal
 
 import com.github.panpf.sketch.SkiaAnimatedImage
-import com.github.panpf.sketch.source.DataSource
 import com.github.panpf.sketch.decode.DecodeResult
 import com.github.panpf.sketch.decode.Decoder
 import com.github.panpf.sketch.decode.ImageInfo
 import com.github.panpf.sketch.request.animationEndCallback
 import com.github.panpf.sketch.request.animationStartCallback
+import com.github.panpf.sketch.request.cacheDecodeTimeoutFrame
 import com.github.panpf.sketch.request.internal.RequestContext
 import com.github.panpf.sketch.request.repeatCount
+import com.github.panpf.sketch.source.DataSource
 import okio.buffer
 import org.jetbrains.skia.Codec
 import org.jetbrains.skia.Data
@@ -31,11 +32,13 @@ open class SkiaAnimatedDecoder(
         // TODO not support resize
         val request = requestContext.request
         val repeatCount = request.repeatCount
+        val cacheDecodeTimeoutFrame = request.cacheDecodeTimeoutFrame == true
         val resize = requestContext.computeResize(imageInfo.size)
         DecodeResult(
             image = SkiaAnimatedImage(
                 codec = codec,
                 repeatCount = repeatCount,
+                cacheDecodeTimeoutFrame = cacheDecodeTimeoutFrame,
                 animationStartCallback = request.animationStartCallback,
                 animationEndCallback = request.animationEndCallback
             ),