+++ sketch-animated/src/nonAndroidTest/kotlin/com/github/panpf/sketch/animated/nonandroid/test/WebpSkiaAnimatedDecoderTest.kt
@@ -1,5 +1,0 @@
-package com.github.panpf.sketch.animated.nonandroid.test
-
-class WebpSkiaAnimatedDecoderTest {
-    // TODO test
-}