+++ sketch-animated/src/nonAndroidTest/kotlin/com/github/panpf/sketch/animated/nonandroid/test/request/AnimatedExtensionsNonAndroidTest.kt
@@ -1,0 +1,5 @@
+package com.github.panpf.sketch.animated.nonandroid.test.request
+
+class AnimatedExtensionsNonAndroidTest {
+    // TODO test
+}